+++ SwiftUI_Graphics/Radar/AxisView.swift
@@ -8,12 +8,6 @@
 
 import SwiftUI
 
-struct AxisComponent: Hashable {
-    
-    var subtitle: String
-    var scale: ClosedRange<Double>
-}
-
 struct AxisView: View {
     
     var color: Color
@@ -80,23 +74,6 @@ struct SubtitleInGeoProxyModifier: ViewModifier {
         content.position(geoProxy.center().findPointIn(radius: 175, radians: self.radians))
     }
     
-}
-
-
-extension CGPoint {
-    static func +(p1:CGPoint, p2:CGPoint) -> CGPoint {
-        return CGPoint(x: p1.x + p2.x, y: p1.y+p2.y)
-    }
-    
-    func findPointIn(radius: Double, radians: Double) -> CGPoint{
-        return self + CGPoint(x: radius * cos(radians), y: radius * sin(radians))
-    }
-    
-}
-extension GeometryProxy {
-    func center() -> CGPoint {
-        return CGPoint(x: self.frame(in: .global).midX, y: self.frame(in: .global).midY)
-    }
 }
 
 struct AxisView_Previews: PreviewProvider {
