--- SwiftUI_Graphics/Radar/GraphView.swift
@@ -1,14 +1,0 @@
-//
-//  GraphView.swift
-//  SwiftUI_Graphics
-//
-//  Created by Paula Leite on 12/06/20.
-//  Copyright © 2020 Luiza Fattori. All rights reserved.
-//
-
-import SwiftUI
-
-struct GraphElement {
-    var subtitle: String
-    var values: [Double]
-}
